v.0.2.3.b2
fixed ominous bug related to the fact that ApiClient.interceptors.request.use starts to work only after reloading page where the interceptor is imported, before that it won't be possible to attach an authorization header with a token (despite the fact that the token has already been received), you can bypass the bug if you use local storage for the access token.

Perhaps the reason is that the interceptor isn't updated due to the fact that the token is in useRef, but even if you change the token storage type to useState, the problem won't be solved - the interceptor will be re-rendered, but interceptors.request will also not be applied.

The bug was bypassed by double importing the interceptor - in AppRouter and the component where it is needed